New rectangle layer 1. Refactored code because rectangle layer uses majority of the grid layer's code. 1. Implemented rectangle layer. 1. Bumped ALS version.
