+++ SynthLineLayer/SynthLineWizard.js
@@ -1,7 +1,3 @@
-L.ALS.SynthLineWizard = L.ALS.Wizard.extend({
+L.ALS.SynthLineWizard = L.ALS.EmptyWizard.extend({
 	displayName: "lineLayerName",
-	initialize: function () {
-		L.ALS.Wizard.prototype.initialize.call(this);
-		this.addWidget(new L.ALS.Widgets.SimpleLabel("lbl", "lineLayerWizardLabel").setStyle("message"));
-	}
 });
