+++ SynthRectangleLayer/SynthRectangleWizard.js
@@ -1,0 +1,3 @@
+L.ALS.SynthRectangleWizard = L.ALS.EmptyWizard.extend({
+	displayName: "rectangleLayerName"
+});
